job/producer/requirements.txt
python-dotenv

kafka-python==2.0.5


